Move Public API to EventBroker
